requirements.txt
fastapi
uvicorn
pyjwt
sqlalchemy
pydantic
passlib[bcrypt]
django-environ
pytest
httpx
apscheduler